Remove apply to one case with percentage discount
Apply to one with percentage discount is more complicated case with
product base discounts. On which product we should apply it?
On first, last or cheapest?

Now percentage case for products based discounts is limited
to the all value and the apply_to field is not used in this case.